+++ apps/worldgen/src/layer_ops/river_mapping/vector_ops.rs
@@ -3,8 +3,8 @@ use game_data_codec::LayerPack;
 
 #[rustfmt::skip]
 impl RgParams {
-	pub fn vector_within_len(&self, allowed: usize,) -> bool {self.length() <= allowed}
-	pub fn vector_start(&mut self, i: usize, j: usize,) {self.dv.x0 = i; self.dv.y0 = j;}
+	pub fn vector_within_len(&self, allowed: u32,) -> bool {self.length() <= allowed}
+	pub fn vector_start(&mut self, i: u32, j: u32,) {self.dv.x0 = i; self.dv.y0 = j;}
 	
 	pub fn vector_end(&mut self, lp: &mut LayerPack,) {
 		// Decide how to find the end of the river vector.
@@ -15,7 +15,7 @@ impl RgParams {
 		}
 	}
 	
-	pub fn vector_bound(&mut self,allowed: usize,) {
+	pub fn vector_bound(&mut self,allowed: u32,) {
 		// Bound up.
 		if self.dv.x0 >= allowed {self.dv.x0 = allowed - 1;}
 		if self.dv.y0 >= allowed {self.dv.y0 = allowed - 1;}
@@ -28,12 +28,12 @@ impl RgParams {
 		if self.dv.y1 == 0 {self.dv.y1 = 0;}
 	}
 	
-	pub fn length(&self) -> usize {
+	pub fn length(&self) -> u32 {
 		((self.dv.x0 as f32 - self.dv.x1 as f32).powf(2.0) + (self.dv.y0 as f32 - self.dv.y1 as f32).powf(2.0))
-		.powf(0.5) as usize
+		.powf(0.5) as u32
 	}
 	
-	pub fn distance(&self) -> usize {
-		((self.dv.x0 as i32 - self.dv.x1 as i32).abs() + (self.dv.y0 as i32 - self.dv.y1 as i32).abs()) as usize
+	pub fn distance(&self) -> u32 {
+		((self.dv.x0 as i32 - self.dv.x1 as i32).abs() + (self.dv.y0 as i32 - self.dv.y1 as i32).abs()) as u32
 	}
 }
